Moved Windows Device Portal classes out of the editor folder
