+++ Assets/MixedRealityToolkit/BuildAndDeploy/Editor/BuildDeployPreferences.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using MixedRealityToolkit.Build.WindowsDevicePortal.DataStructures;
+using MixedRealityToolkit.Common.WindowsDevicePortal.DataStructures;
 using MixedRealityToolkit.Common.EditorScript;
 using System;
 using System.IO;
