+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal.meta
