+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures.meta
