+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/ActivePowerSchemeInfo.cs.meta
