+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/AdapterInfo.cs.meta
