+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/ApplicationInfo.cs.meta
