+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/BatteryInfo.cs.meta
