+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/DeviceInfo.cs.meta
