+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/DeviceOsInfo.cs.meta
