+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/DevicePortalConnections.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace MixedRealityToolkit.Build.WindowsDevicePortal.DataStructures
+namespace MixedRealityToolkit.Common.WindowsDevicePortal.DataStructures
 {
     [Serializable]
     public class DevicePortalConnections
