+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/FileInfo.cs.meta
