+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/FileList.cs.meta
