+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/InstallStatus.cs.meta
