+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/InstalledApps.cs
@@ -3,11 +3,11 @@
 
 using System;
 
-namespace MixedRealityToolkit.Build.WindowsDevicePortal.DataStructures
+namespace MixedRealityToolkit.Common.WindowsDevicePortal.DataStructures
 {
     [Serializable]
-    public class FileList
+    public class InstalledApps
     {
-        public FileInfo[] Items;
+        public ApplicationInfo[] InstalledPackages;
     }
 }
