+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/InstalledApps.cs.meta
