+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/InterfaceInfo.cs.meta
