+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/IpConfigInfo.cs.meta
