+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/MachineName.cs.meta
