+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/NetworkInterfaces.cs.meta
