+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/NetworkProfileInfo.cs.meta
