+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/PowerStateInfo.cs.meta
