+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/ProcessList.cs
@@ -3,11 +3,11 @@
 
 using System;
 
-namespace MixedRealityToolkit.Build.WindowsDevicePortal.DataStructures
+namespace MixedRealityToolkit.Common.WindowsDevicePortal.DataStructures
 {
     [Serializable]
-    public class NetworkInterfaces
+    public class ProcessList
     {
-        public InterfaceInfo[] Interfaces;
+        public ProcessInfo[] Processes;
     }
 }
