+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DataStructures/WirelessNetworkInfo.cs
@@ -3,7 +3,7 @@
 
 using System;
 
-namespace MixedRealityToolkit.Build.WindowsDevicePortal.DataStructures
+namespace MixedRealityToolkit.Common.WindowsDevicePortal.DataStructures
 {
     [Serializable]
     public class WirelessNetworkInfo
