+++ Assets/MixedRealityToolkit/Common/Scripts/WindowsDevicePortal/DevicePortal.cs.meta
